Move old C06 to new C07
Move the old code, refactor to use file-scoped namespaces, and remove
unused using statements.
