--- 2ndEd/C06/Strategy/src/MySortingMachine/ISortStrategy.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-
-namespace MySortingMachine
-{
-    public interface ISortStrategy
-    {
-        IOrderedEnumerable<string> Sort(IEnumerable<string> input);
-    }
-}
