--- 2ndEd/C06/Strategy/src/MySortingMachine/SortDescendingStrategy.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-
-namespace MySortingMachine
-{
-    public class SortDescendingStrategy : ISortStrategy
-    {
-        public IOrderedEnumerable<string> Sort(IEnumerable<string> input)
-            => input.OrderByDescending(x => x);
-    }
-
-    public class SortDescendingStrategyClassic : ISortStrategy
-    {
-        public IOrderedEnumerable<string> Sort(IEnumerable<string> input)
-        {
-            return input.OrderByDescending(x => x);
-        }
-    }
-}
