+++ C07/AbstractFactory/AbstractVehicleFactory/AbstractFactory.sln
