+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/HighGradeVehicleFactory.cs
