+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/IVehicleFactory.cs
