+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/LowGradeVehicleFactory.cs
