+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/Models/HighGradeBike.cs
