+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/Models/HighGradeCar.cs
