+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/Models/IBike.cs
