+++ C07/AbstractFactory/AbstractVehicleFactory/src/Vehicles/Models/LowGradeBike.cs
