+++ C07/AbstractFactory/AbstractVehicleFactory/test/Vehicles.Tests/AbstractFactoryBaseTestData.cs
