+++ C07/AbstractFactory/AbstractVehicleFactory/test/Vehicles.Tests/AbstractFactoryTest.cs
