+++ C07/AbstractFactory/MiddleEndVehicleFactory/AbstractFactoryExpansion.sln
