+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/HighGradeVehicleFactory.cs
