+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/IVehicleFactory.cs
