+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/LowGradeVehicleFactory.cs
