+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/MiddleEndVehicleFactory.cs
