+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/HighGradeBike.cs
