+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/HighGradeCar.cs
