+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/IBike.cs
