+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/LowGradeBike.cs
