+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/LowGradeCar.cs
