+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/MiddleGradeBike.cs
