+++ C07/AbstractFactory/MiddleEndVehicleFactory/src/Vehicles/Models/MiddleGradeCar.cs
