+++ C07/AbstractFactory/MiddleEndVehicleFactory/test/Vehicles.Tests/AbstractFactoryBaseTestData.cs
