+++ C07/AbstractFactory/MiddleEndVehicleFactory/test/Vehicles.Tests/AbstractFactoryTest.cs
