+++ C07/AbstractFactory/MiddleEndVehicleFactory/test/Vehicles.Tests/Vehicles.Tests.csproj
