+++ C07/Singleton/Singleton.sln
