+++ C07/Singleton/src/Singleton/MyAmbientContext.cs
