+++ C07/Singleton/src/Singleton/MySimpleSingleton.cs
