+++ C07/Singleton/src/Singleton/MySingleton.cs
