+++ C07/Singleton/src/Singleton/MySingletonWithLock.cs
