+++ C07/Singleton/src/Singleton/Singleton.csproj
