+++ C07/Singleton/test/Singleton.Tests/MyAmbientContextTest.cs
