+++ C07/Singleton/test/Singleton.Tests/MySimpleSingletonTest.cs
