+++ C07/Singleton/test/Singleton.Tests/MySingletonTest.cs
