+++ C07/Singleton/test/Singleton.Tests/MySingletonWithLockTest.cs
