+++ C07/Singleton/test/Singleton.Tests/Singleton.Tests.csproj
