+++ C07/Strategy/Strategy.sln
