+++ C07/Strategy/src/MyConsumerApp/MyConsumerApp.csproj
