+++ C07/Strategy/src/MySortingMachine/ISortStrategy.cs
@@ -1,0 +1,6 @@
+﻿namespace MySortingMachine;
+
+public interface ISortStrategy
+{
+    IOrderedEnumerable<string> Sort(IEnumerable<string> input);
+}
