+++ C07/Strategy/src/MySortingMachine/MySortingMachine.csproj
