+++ C07/Strategy/src/MySortingMachine/SortAscendingStrategy.cs
@@ -1,0 +1,7 @@
+﻿namespace MySortingMachine;
+
+public class SortAscendingStrategy : ISortStrategy
+{
+    public IOrderedEnumerable<string> Sort(IEnumerable<string> input) 
+        => input.OrderBy(x => x);
+}
