+++ C07/Strategy/src/MySortingMachine/SortDescendingStrategy.cs
@@ -1,0 +1,15 @@
+﻿namespace MySortingMachine;
+
+public class SortDescendingStrategy : ISortStrategy
+{
+    public IOrderedEnumerable<string> Sort(IEnumerable<string> input)
+        => input.OrderByDescending(x => x);
+}
+
+public class SortDescendingStrategyClassic : ISortStrategy
+{
+    public IOrderedEnumerable<string> Sort(IEnumerable<string> input)
+    {
+        return input.OrderByDescending(x => x);
+    }
+}
